+++ src/main/java/de/fh_zwickau/informatik/sensor/model/instances/dummydevice/DummyDeviceParams.java
@@ -1,0 +1,49 @@
+/**
+ * Copyright (C) 2016 by Software-Systementwicklung Zwickau Research Group
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
+package de.fh_zwickau.informatik.sensor.model.instances.dummydevice;
+
+import org.apache.commons.lang3.builder.ToStringBuilder;
+
+import com.google.gson.annotations.SerializedName;
+
+/**
+ * The {@link DummyDeviceParams} represents a part of dummy device configuration. See also:
+ * http://docs.zwayhomeautomation.apiary.io/#
+ *
+ * @author Patrick Hecker - Initial contribution
+ */
+public class DummyDeviceParams {
+    @SerializedName("deviceType")
+    private String mDeviceType;
+
+    /**
+     * Instantiate a the params part of dummy device configuration with default values.
+     */
+    public DummyDeviceParams() {
+        mDeviceType = "";
+    }
+
+    public String getDeviceType() {
+        return mDeviceType;
+    }
+
+    public void setDeviceType(String deviceType) {
+        this.mDeviceType = deviceType;
+    }
+
+    /*
+     * (non-Javadoc)
+     *
+     * @see java.lang.Object#toString()
+     */
+    @Override
+    public String toString() {
+        return new ToStringBuilder(this).append("deviceType", mDeviceType).toString();
+    }
+}